+++ tests/testthat/test-calc_indicator.R
@@ -87,9 +87,6 @@ test_that("calc_indicator works", {
     c(2005.175, 2004.664, 2002.985, 1968.824, 1959.262, 1955.029),
     tolerance = 1e-3
   )
-  expect_warning(
-    calc_indicators(portfolio, "treecover")
-  )
 })
 
 test_that("Parallelization works", {